1. 文件为0，不允许上传 2. 上传相同文件不新建文件，建立引用关系 3. 删除时如果存在该文件被引用，重新建立引用、被引用关系，再删除

+++ src/main/java/com/hxqh/filemanager/service/FileServiceImpl.java
@@ -7,6 +7,7 @@ import com.hxqh.filemanager.model.User;
 import com.hxqh.filemanager.model.assist.FileDto;
 import com.hxqh.filemanager.model.assist.FileInfo;
 import com.hxqh.filemanager.model.assist.FileVersionDto;
+import com.hxqh.filemanager.model.assist.Refer;
 import com.hxqh.filemanager.repository.FileRepository;
 import com.hxqh.filemanager.repository.FileVersionRepository;
 import com.hxqh.filemanager.repository.UserRepository;
@@ -133,6 +134,25 @@ public class FileServiceImpl implements FileService {
         return fileVersionRepository.findByFileversionid(fid);
     }
 
+    @Transactional(readOnly = true, rollbackFor = Exception.class)
+    @Override
+    public boolean hasSameFIle(MultipartFile files, Integer fileid) throws Exception {
+        String md5String = Md5Utils.getFileMD5String(files.getBytes());
+        TbFile file = fileRepository.findByFileid(fileid);
+        List<TbFileVersion> fileVersionByMd5 = fileVersionRepository.findByMd5(md5String);
+
+        if (md5String.equals(file.getMd5())) {
+            return true;
+        }
+        for (int i = 0; i < fileVersionByMd5.size(); i++) {
+            TbFileVersion fileVersion = fileVersionByMd5.get(i);
+            if (md5String.equals(fileVersion.getMd5())) {
+                return true;
+            }
+        }
+        return false;
+    }
+
 
     @Transactional(readOnly = true, rollbackFor = Exception.class)
     @Override
@@ -142,6 +162,9 @@ public class FileServiceImpl implements FileService {
 
     /**
      * 保存文件
+     * <p>
+     * 先判断tb_file tb_fileverson是否存在相同的md5
+     * 相同的md5文件不新建文件，设置引用关系
      *
      * @param file     文件
      * @param fileInfo 文件信息
@@ -149,29 +172,35 @@ public class FileServiceImpl implements FileService {
     @Transactional(rollbackFor = Exception.class)
     @Override
     public void saveFile(MultipartFile file, FileInfo fileInfo) throws Exception {
+        Refer refer = null;
         String filePath, savePath;
-        String md5String = null;
+        String md5String = Md5Utils.getFileMD5String(file.getBytes());
 
-        File f = new File(uploadPath + "/" + DateUtils.getTodayMonth());
-        if (!f.exists()) {
-            f.mkdirs();
-        }
-        String extensionName = file.getOriginalFilename().split("\\.")[1];
-        savePath = "/" + DateUtils.getTodayMonth() + "/" + DateUtils.getTodayTime() + "_" + UUID.randomUUID()
-                + "." + extensionName;
+        List<TbFile> fileByMd5 = fileRepository.findByMd5(md5String);
+        List<TbFileVersion> fileVersionByMd5 = fileVersionRepository.findByMd5(md5String);
+
+        // 存储路径
+        mkdirStorePath();
+
+        // 生成随机文件名称
+        savePath = generateFileName(file);
 
         if (file.getOriginalFilename() != null && file.getSize() > 0) {
-
-            filePath = uploadPath + savePath;
-            FileOutputStream outputStream;
-            try {
-                outputStream = new FileOutputStream(new File(filePath));
-                md5String = Md5Utils.getFileMD5String(file.getBytes());
-                outputStream.write(file.getBytes());
-                outputStream.flush();
-                outputStream.close();
-            } catch (IOException e) {
-                logger.error(e.getMessage());
+            // 保存至文件系统
+            if (fileByMd5.size() == 0 && fileVersionByMd5.size() == 0) {
+                filePath = uploadPath + savePath;
+                FileOutputStream outputStream;
+                try {
+                    outputStream = new FileOutputStream(new File(filePath));
+                    outputStream.write(file.getBytes());
+                    outputStream.flush();
+                    outputStream.close();
+                } catch (IOException e) {
+                    logger.error(e.getMessage());
+                }
+            } else {
+                refer = getSavePath(fileByMd5, fileVersionByMd5);
+                savePath = refer.getSavePath();
             }
 
             if (null == fileInfo.getFileid()) {
@@ -179,8 +208,14 @@ public class FileServiceImpl implements FileService {
                 TbFile tbFile = new TbFile();
                 setFileProperties(file, fileInfo, savePath, tbFile);
                 tbFile.setMd5(md5String);
+                tbFile.setFilename(file.getOriginalFilename().split("\\.")[0]);
+                tbFile.setExtensionname(file.getOriginalFilename().split("\\.")[1]);
+                if (null != refer) {
+                    BeanUtils.copyProperties(refer, tbFile);
+                }
                 fileRepository.save(tbFile);
             } else {
+                // 保存文件版本信息
                 TbFile tbFile = fileRepository.findByFileid(fileInfo.getFileid());
                 TbFileVersion fileVersion = new TbFileVersion();
                 BeanUtils.copyProperties(tbFile, fileVersion);
@@ -193,10 +228,39 @@ public class FileServiceImpl implements FileService {
                 BeanUtils.copyProperties(newVersion, tbFile);
                 fileVersion.setTbFile(tbFile);
                 fileVersion.setMd5(md5String);
+                if (null != refer) {
+                    BeanUtils.copyProperties(refer, fileVersion);
+                }
                 fileVersionRepository.save(fileVersion);
             }
         }
 
+    }
+
+    private Refer getSavePath(List<TbFile> fileByMd5, List<TbFileVersion> fileVersionByMd5) {
+        Refer refer = null;
+        if (fileByMd5.size() > 0) {
+            refer = new Refer(IConstants.FILE_REFER, fileByMd5.get(0).getFileid(), fileByMd5.get(0).getFilepath());
+        }
+        if (fileVersionByMd5.size() > 0) {
+            refer = new Refer(IConstants.VERSION_REFER, fileVersionByMd5.get(0).getFileversionid(), fileVersionByMd5.get(0).getFilepath());
+        }
+        return refer;
+    }
+
+    private String generateFileName(MultipartFile file) {
+        String savePath;
+        String extensionName = file.getOriginalFilename().split("\\.")[1];
+        savePath = "/" + DateUtils.getTodayMonth() + "/" + DateUtils.getTodayTime() + "_" + UUID.randomUUID()
+                + "." + extensionName;
+        return savePath;
+    }
+
+    private void mkdirStorePath() {
+        File f = new File(uploadPath + "/" + DateUtils.getTodayMonth());
+        if (!f.exists()) {
+            f.mkdirs();
+        }
     }
 
     private TbFile setFileProperties(MultipartFile file, FileInfo fileInfo, String savePath, TbFile tbFile) {
@@ -215,11 +279,43 @@ public class FileServiceImpl implements FileService {
     @Transactional(rollbackFor = Exception.class)
     @Override
     public void deleteFile(FileInfo fileInfo) {
+
         if (null != fileInfo.getFileid()) {
             // 文件系统删除
             TbFile file = fileRepository.findByFileid(fileInfo.getFileid());
             String filePath = uploadPath + file.getFilepath();
-            FileUtils.deleteFile(filePath);
+
+            // 先查询主记录是否存在被引用情况
+            List<TbFile> fileList = fileRepository.findByRefertabAndReferid(IConstants.FILE_REFER, file.getFileid());
+            // 选取一条设置
+            for (int i = 0; i < fileList.size(); i++) {
+                if (0 == i) {
+                    fileList.get(0).setReferid(null);
+                    fileList.get(0).setRefertab(null);
+                    fileRepository.save(fileList.get(0));
+                } else {
+                    fileList.get(i).setReferid(fileList.get(0).getFileid());
+                    fileRepository.save(fileList.get(i));
+                }
+            }
+
+            List<TbFileVersion> fileVersionList = fileVersionRepository.findByRefertabAndReferid(IConstants.FILE_REFER, file.getFileid());
+            // 选取一条设置
+            for (int i = 0; i < fileVersionList.size(); i++) {
+                TbFileVersion fileVersion = fileVersionList.get(i);
+                if (i == 0) {
+                    fileVersion.setReferid(null);
+                    fileVersion.setRefertab(null);
+                } else {
+                    fileVersion.setReferid(fileVersionList.get(0).getFileid());
+                }
+                fileVersionRepository.save(fileVersion);
+            }
+
+
+            if (fileList.size() == 0 && fileVersionList.size() == 0) {
+                FileUtils.deleteFile(filePath);
+            }
 
             List<TbFileVersion> fileVersions = file.getTbFileVersions();
             for (int i = 0; i < fileVersions.size(); i++) {
